+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased][]
 
+### Changed
+
+- Use lower browser version
+
 ## [2.0.0][] - 2025-11-06
 
 ### Changed